Modified order of variable addition
The variable price/sqfeet was added earlier in order to also be added to previous data set. The goal is to make a heat map for the entire USA and one that focuses on massacheussets.

Next goal is to make heat map of average price/sqfeet per year in each zipcode